Removed websocket config and created a dedicated interface; refactor websocket in  a new package and add new IWebSocketSettings interface to configure settings;
