+++ StreamNode/Services/Settings/ISettings.cs
@@ -1,15 +1,11 @@
 ﻿using System.ComponentModel;
 using StreamNodeEngine.Engine.Services.Obs;
+using StreamNodeEngine.Engine.Services.WebSocket;
 
 namespace StreamNode.Services.Settings
 {
-    public interface ISettings : IObsSettings
+    public interface ISettings : IObsSettings, IWebSocketSettings
     {
-        /**
-         * Interface for Config Library to Read Settings
-         */
-        [DefaultValue(8189)]
-        int WebSocketPort { get; set; }
 
         [DefaultValue("*")]
         string HttpServerIp { get; set; }
